seprate voxel dispatcher
